refactor: move timeDifference key to State type (#200)
This moves the `timeDifference` key from `StrategyOptions` to  the`State` type.
This is because `timeDifference` is only used internally in the state
object and not passed in by the user in the options parameter to
auth-app.

The types were not failing previously because the `State` type extends
the `StrategyOptions` type. However, `State` is the correct place for
this key to live.